+++ packages/apollo-links/src/auth-link/authLink.ts
@@ -4,37 +4,53 @@
 import { ApolloLink, FetchResult, NextLink, Observable, Operation } from '@apollo/client/core';
 import { Subscription } from 'zen-observable-ts';
 
-import { isTokenExpired, POST, requestAuthToken } from './authHelper';
+import { isTokenExpired, requestAuthToken } from './authHelper';
 import { Message } from './eip712';
+import { Logger } from '../logger';
 
-export interface AuthOptions extends Message {
-  authUrl: string;   // the url for geting token
-  chainId: number;   // chainId for the network
-  sk?: string;       // `sk` of the consumer or corresponding controller account
+interface AuthOptions extends Message {
+  indexerUrl: string; // indexer url
+  chainId: number; // chainId for the network
+  sk: string; // `sk` of the consumer or corresponding controller account
 }
 
 export class AuthLink extends ApolloLink {
   private _options: AuthOptions;
+  private _logger: Logger;
   private _token: string;
 
-  constructor(options: AuthOptions) {
+  constructor(options: AuthOptions, logger: Logger) {
     super();
     this._options = options;
+    this._logger = logger;
     this._token = '';
   }
 
   override request(operation: Operation, forward?: NextLink): Observable<FetchResult> | null {
     if (!forward) return null;
 
-    return new Observable<FetchResult>(observer => {
+    return new Observable<FetchResult>((observer) => {
       let sub: Subscription;
-      this.requestToken().then((token) => {
-        operation.setContext({ headers: { authorization: `Bearer ${token}` } }); 
-        sub = forward(operation).subscribe(observer);
-      });
+      this.getUrlAndToken()
+        .then((data) => {
+          if (data) {
+            const { token, url } = data;
+            const headers = { authorization: `Bearer ${token}` };
+            operation.setContext({ url, headers });
+          }
+        })
+        .catch((error) => observer.error(error))
+        .finally(() => {
+          sub = forward(operation).subscribe(observer);
+        });
 
-      return () => sub.unsubscribe();
+      return () => sub?.unsubscribe();
     });
+  }
+
+  get queryEndpoint() {
+    const url = new URL('/query', this._options.indexerUrl);
+    return url.toString();
   }
 
   private generateMessage() {
@@ -43,21 +59,18 @@ export class AuthLink extends ApolloLink {
     return { indexer, consumer, agreement, deploymentId, timestamp };
   }
 
-  private async requestToken(): Promise<string> {
-    if (!isTokenExpired(this._token)) return this._token;
+  private async getUrlAndToken(): Promise<{ url: string; token: string } | undefined> {
+    if (!isTokenExpired(this._token)) return { token: this._token, url: this.queryEndpoint };
 
-    const { indexer, deploymentId, sk, chainId, authUrl } = this._options;
+    const { sk, chainId, agreement, indexerUrl } = this._options;
 
-    if (!sk) {
-      const host = authUrl?.trim().replace(/\/+$/, '');
-      const res = await POST<{ token: string }>(`${host}/token`, { deploymentId, indexer });
-      this._token = res.token;
-      return this._token;
-    } 
+    if (!chainId || !agreement) throw new Error('chainId and agreement are required');
 
     const message = this.generateMessage();
-    this._token = await requestAuthToken(authUrl, message, sk, chainId)
+    const tokenUrl = new URL('/token', indexerUrl);
+    const authToken = await requestAuthToken(tokenUrl.toString(), message, sk, chainId);
+    this._token = authToken;
 
-    return this._token;
+    return { token: authToken, url: this.queryEndpoint };
   }
 }