+++ test/networkClient.test.ts
@@ -1,15 +1,14 @@
 // Copyright 2020-2022 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { NetworkClient } from '../packages/network-clients';
-import { SQNetworks } from '../packages/network-clients';
+import { NetworkClient, SQNetworks } from '../packages/network-clients/src';
 
 const TEST_INDEXER = '0xCef192586b70e3Fc2FAD76Dd1D77983a30d38D04';
 
 describe.only('network client', () => {
   let client: NetworkClient;
-  beforeAll(async () => {
-    client = await NetworkClient.create(SQNetworks.TESTNET);
+  beforeAll(() => {
+    client = NetworkClient.create(SQNetworks.TESTNET);
   });
 
   it('can get indexer detail', async () => {